change component to app and add sample component to the newly generated app
